modified version 1

+++ KnowledgeCheck1_Calculator/Calculator.cs
@@ -2,22 +2,22 @@ namespace KnowledgeCheck1_Calculator
 {
     public class Calculator
     {
-        public int Add(int first, int second)
+        public double Add(double first, double second)
         {
             return first + second;
         }
 
-        public int Subtract(int first, int second)
+        public double Subtract(double first, double second)
         {
             return first - second;
         }
 
-        public int Multiply(int first, int second)
+        public double Multiply(double first, double second)
         {
             return first * second;
         }
 
-        public double Divide(int first, int second)
+        public double Divide(double first, double second)
         {
             if(second == 0)
             {
